+++ src/GSS.Authorization.OAuth2.HttpClient/ServiceCollectionExtensions.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Caching.Memory;
-using Microsoft.Extensions.Logging;
 
 namespace GSS.Authorization.OAuth2
 {
@@ -41,8 +40,7 @@ namespace GSS.Authorization.OAuth2
                 .AddHttpClient<TClient>()
                 .AddHttpMessageHandler(resolver => new OAuth2HttpHandler(
                     resolver.GetRequiredService<TAuthorizer>(),
-                    resolver.GetRequiredService<IMemoryCache>(),
-                    resolver.GetRequiredService<ILoggerFactory>()));
+                    resolver.GetRequiredService<IMemoryCache>()));
         }
 
         /// <summary>
@@ -76,8 +74,7 @@ namespace GSS.Authorization.OAuth2
                 .AddHttpClient(name)
                 .AddHttpMessageHandler(resolver => new OAuth2HttpHandler(
                     resolver.GetRequiredService<TAuthorizer>(),
-                    resolver.GetRequiredService<IMemoryCache>(),
-                    resolver.GetRequiredService<ILoggerFactory>()));
+                    resolver.GetRequiredService<IMemoryCache>()));
         }
 
         /// <summary>